Move CheckResult to model directory
