+++ src/kathara_lab_checker/checks/DaemonCheck.py
@@ -2,7 +2,7 @@ from Kathara.exceptions import MachineNotFoundError
 
 from ..utils import get_output
 from .AbstractCheck import AbstractCheck
-from .CheckResult import CheckResult
+from ..model.CheckResult import CheckResult
 
 
 class DaemonCheck(AbstractCheck):
