+++ src/kathara_lab_checker/checks/IPv6EnabledCheck.py
@@ -1,7 +1,7 @@
 from Kathara.exceptions import MachineNotFoundError
 
 from .AbstractCheck import AbstractCheck
-from .CheckResult import CheckResult
+from ..model.CheckResult import CheckResult
 
 
 class IPv6EnabledCheck(AbstractCheck):
