+++ src/kathara_lab_checker/checks/ReachabilityCheck.py
@@ -2,7 +2,7 @@ import jc
 
 from ..utils import get_output
 from .AbstractCheck import AbstractCheck
-from .CheckResult import CheckResult
+from ..model.CheckResult import CheckResult
 
 
 class ReachabilityCheck(AbstractCheck):
