+++ src/kathara_lab_checker/checks/StartupExistenceCheck.py
@@ -1,5 +1,5 @@
 from .AbstractCheck import AbstractCheck
-from .CheckResult import CheckResult
+from ..model.CheckResult import CheckResult
 
 
 class StartupExistenceCheck(AbstractCheck):
