+++ src/kathara_lab_checker/checks/applications/dns/DNSAuthorityCheck.py
@@ -4,7 +4,7 @@ import jc
 from Kathara.exceptions import MachineNotRunningError
 
 from ...AbstractCheck import AbstractCheck
-from ...CheckResult import CheckResult
+from ....model.CheckResult import CheckResult
 from ....utils import get_output, find_lines_with_string, find_device_name_from_ip
