+++ src/kathara_lab_checker/checks/applications/dns/DNSRecordCheck.py
@@ -1,5 +1,5 @@
 from ...AbstractCheck import AbstractCheck
-from ...CheckResult import CheckResult
+from ....model.CheckResult import CheckResult
 from ....utils import get_output
