+++ src/kathara_lab_checker/checks/protocols/ProtocolRedistributionCheck.py
@@ -1,7 +1,7 @@
 import re
 
 from ..AbstractCheck import AbstractCheck
-from ..CheckResult import CheckResult
+from ...model.CheckResult import CheckResult
 from ...utils import get_output
