+++ src/kathara_lab_checker/checks/protocols/bgp/BGPPeeringCheck.py
@@ -3,7 +3,7 @@ import json
 from Kathara.exceptions import MachineNotRunningError
 
 from ...AbstractCheck import AbstractCheck
-from ...CheckResult import CheckResult
+from ....model.CheckResult import CheckResult
 
 
 class BGPPeeringCheck(AbstractCheck):
