+++ src/kathara_lab_checker/model/CheckResult.py
